adds placeholder adapting
placeholder changes based on the selected route
